reorganize for general evaluator
